Problem: building and running complex systems
Modern software systems are typically complex and present significant overhead.
On the architecture spectrum from monolith to SOA, this complexity manifests
through a host of issues.

For example, the data is often copied and transferred over the network to reach
the code that can process it. A simple example would be an application server
sending a query to the database server, getting results, and making further
changes in the database. A more complex example would involve extending this
pipeline with a search engine like Elasticsearch and shuttling data between the
database server and the search engine to get the data indexed. Finally, a
typical example in an SOA architecture would be a service sending data to
another service for further processing; this is particularly common in polyglot
environments where direct integration of service components is less feasible.

Relative closedness of ecosystems of individual stacks results in significant
redundancies and inefficiencies. For example, NLP analysis in Elasticsearch
can't be easily substituted with the one in spaCy (a Python-based NLP toolkit),
so often, the efforts to maintain coherent NLP across the ML and indexing
results in doing the work multiple times, both on the development side as well
as processing.

Another manifestation of complexity comes from observing the management of
system change rollouts. Having multiple components with their own way of
operational management, requirements, etc., creates a rigamarole of a
deployment pipeline. It often leads to increased overhead by introducing
isolation primitives (such as containers).

These and other issues may indicate systemic problems in how we design our
application for the requirements and scale we have these days.

Hypothesis
==========

Software development and bundling practices stem from the paradigm that hinges
on a single-computer view of the world. Every program has a foundation in the
form of some kind of operating system, which handles such resources as memory,
files, isolation, networking, etc. One can visualize this as a pyramid where a
program is at the top and services provided by the OS are below it. However, we
can observe that the vector of this functionality is directed towards the
management of resources of the computer that the program is running on, so it
is physically directed "inwards". This approach makes sense if your entire
system works on that computer. However, with our systems' increased complexity
and scale, this is no longer true.

What if we replaced that foundation and most of the services of such an OS were
directed outwards and provided functionality that is most useful for running
systems that span the network, can be partially rolled out, process data in
complex pipelines, etc.?

Observation: Code is smaller than data
======================================

One big issue we can observe is that the amount of data our system handle is
vastly larger than the amount of code that handles it. So why do we send the
data to the code, then? Wouldn't it make sense to send code to the data and
process it there?

A gained momentum of a virtual ISA and ABI that WebAssembly offers (while
acknowledging its flaws and immaturity) allows us to reimagine how data
processing can be done. Sandboxed code deployed across the hardware resources
that manage the data will enable us to process data in place without
serializing and transferring that data back and forth.

What if the OS, as mentioned above, provided building blocks for managing and
processing data, starting from components like B-Trees and growing into more
sophisticated data management primitives? With more sophisticated and
cluster-away scheduling, this can be a foundation for algorithms that naturally
scale across the system.

Solution: start building a micro-OS (ParaOS)

I've made a few early prototypes and my current plan is as follow.

1) Document features and APIs that ParaOS should provide
2) Implement and test these features APIs in emulation (where possible)
3) Implement a WebAssembly -> Native compiler
4) Implement micro-OS itself

Some of these efforts can be done simultaneously, without depending on
each other. The goal is to get something working as soon as possible.

This commit sets up documentation/API framework for (1). Other things
will follow.