Problem: passing of parameters
They are done through locals and the current code does not do that
and instead puts them on the stack

Solution: rewrite parameter collection

Also, while we're at it, let's validate the module to enforce stack
height as it is going to be important down the road. The current method
of enforcement is a bit heavy (as it re-processes the module twice) but
it does the job for now.