Problem: determining the stack needs of functions
I am researching the best way to implement asynchronous operations
throughout ParaOS. It all hinges on its ability to build out certain
primitives that would allow functions to be able to resume their
execution from certain points. A coroutine type of model.

With this model there's a need to know the stack impact of a function to
be able to lay it out for further execution resumption.

Solution: collect max function height

While this is not a perfect indicator of memory required (certain stack
values can be smaller than a word, for example), it is an indicator
nevertheless as it can be used as a starting point.

Stack -> register optimizations between yield boundries that don't leak
across those boundaries can be subtracted from the effective lay out of
the stack.